Lots of changes! -Support for up to 16 bit per channel, gamma corrected colors -More plugins -Changed settings to work with XBee and send 12-bit color data -Smoothed out some of the feature detectors (low passed) to make the -effects look less flickery

+++ SoundExpressor/src/Visualizors/Ambiance.java
@@ -57,7 +57,7 @@ public class Ambiance extends Visualizer {
 		Color c0, c1, c2, c3;
 		
 		
-		double symLight = -(double) NUM_FRONT_RGB_LIGHTS / 2 - 0.5;
+		double symLight = -NUM_FRONT_RGB_LIGHTS / 2.0 - 0.5;
 		double a = 1.0;
 		
 		for(int light = 0; light < NUM_FRONT_RGB_LIGHTS; light++) {
